+++ NEWS.md
@@ -2,6 +2,10 @@ Version 0.6.5
  - implement: export as rtf / export as html
  - fix: the contents copied/exported are not correctly syntax colored
  - fix: stop execution if the source file is not compiled and user choose not to compile it
+ - fix: not correctly parse gdb's output
+ - fix: path not correctly setted for the debugger process
+ - fix: indent line not correctly drawed
+
 
 Version 0.6.4
  - fix: code completion popup not show after '->' inputted